% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{chargeRange}
\alias{chargeRange}
\alias{chargeRange,msgfPar-method}
\alias{chargeRange<-}
\alias{chargeRange<-,msgfPar,msgfParChargeRange-method}
\alias{chargeRange<-,msgfPar,numeric-method}
\title{Get and set the charge range in msgfPar objects}
\usage{
chargeRange(object)

chargeRange(object) <- value

\S4method{chargeRange}{msgfPar}(object)

\S4method{chargeRange}{msgfPar,numeric}(object) <- value

\S4method{chargeRange}{msgfPar,msgfParChargeRange}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either a numeric vector of length 2 or an msgfParChargeRange
object}
}
\value{
In case of the getter a numeric vector with the named elements 'min'
and 'max'
}
\description{
These functions allow you to retrieve and set the charge range in the msgfPar
object of interest
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the charge range

\item \code{object = msgfPar,value = numeric}: Set the charge range using lower and upper bounds

\item \code{object = msgfPar,value = msgfParChargeRange}: Set the charge range using a dedicated
msgfParChargeRange object
}}
\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
chargeRange(parameters) <- c(2, 4)
chargeRange(parameters)
}
\seealso{
Other msgfPar-getter_setter: \code{\link{db}},
  \code{\link{db,msgfPar-method}}, \code{\link{db<-}},
  \code{\link{db<-,msgfPar,character-method}};
  \code{\link{enzyme}},
  \code{\link{enzyme,msgfPar-method}},
  \code{\link{enzyme<-}},
  \code{\link{enzyme<-,msgfPar,character-method}},
  \code{\link{enzyme<-,msgfPar,msgfParEnzyme-method}},
  \code{\link{enzyme<-,msgfPar,numeric-method}};
  \code{\link{fragmentation}},
  \code{\link{fragmentation,msgfPar-method}},
  \code{\link{fragmentation<-}},
  \code{\link{fragmentation<-,msgfPar,character-method}},
  \code{\link{fragmentation<-,msgfPar,msgfParFragmentation-method}},
  \code{\link{fragmentation<-,msgfPar,numeric-method}};
  \code{\link{instrument}},
  \code{\link{instrument,msgfPar-method}},
  \code{\link{instrument<-}},
  \code{\link{instrument<-,msgfPar,character-method}},
  \code{\link{instrument<-,msgfPar,msgfParInstrument-method}},
  \code{\link{instrument<-,msgfPar,numeric-method}};
  \code{\link{isotopeError}},
  \code{\link{isotopeError,msgfPar-method}},
  \code{\link{isotopeError<-}},
  \code{\link{isotopeError<-,msgfPar,msgfParIsotopeError-method}},
  \code{\link{isotopeError<-,msgfPar,numeric-method}};
  \code{\link{lengthRange}},
  \code{\link{lengthRange,msgfPar-method}},
  \code{\link{lengthRange<-}},
  \code{\link{lengthRange<-,msgfPar,msgfParLengthRange-method}},
  \code{\link{lengthRange<-,msgfPar,numeric-method}};
  \code{\link{matches}},
  \code{\link{matches,msgfPar-method}},
  \code{\link{matches<-}},
  \code{\link{matches<-,msgfPar,msgfParMatches-method}},
  \code{\link{matches<-,msgfPar,numeric-method}};
  \code{\link{mods}}, \code{\link{mods,msgfPar-method}},
  \code{\link{mods<-}},
  \code{\link{mods<-,msgfPar,msgfParModificationList-method}},
  \code{\link{nMod}}, \code{\link{nMod,msgfPar-method}},
  \code{\link{nMod<-}},
  \code{\link{nMod<-,msgfPar,numeric-method}};
  \code{\link{ntt}}, \code{\link{ntt,msgfPar-method}},
  \code{\link{ntt<-}},
  \code{\link{ntt<-,msgfPar,msgfParNtt-method}},
  \code{\link{ntt<-,msgfPar,numeric-method}};
  \code{\link{protocol}},
  \code{\link{protocol,msgfPar-method}},
  \code{\link{protocol<-}},
  \code{\link{protocol<-,msgfPar,character-method}},
  \code{\link{protocol<-,msgfPar,msgfParProtocol-method}},
  \code{\link{protocol<-,msgfPar,numeric-method}};
  \code{\link{tda}}, \code{\link{tda,msgfPar-method}},
  \code{\link{tda<-}},
  \code{\link{tda<-,msgfPar,logical-method}},
  \code{\link{tda<-,msgfPar,msgfParTda-method}};
  \code{\link{tolerance}},
  \code{\link{tolerance,msgfPar-method}},
  \code{\link{tolerance<-}},
  \code{\link{tolerance<-,msgfPar,character-method}},
  \code{\link{tolerance<-,msgfPar,msgfParTolerance-method}},
  \code{\link{toleranceRange}},
  \code{\link{toleranceRange,msgfPar-method}},
  \code{\link{toleranceRange<-}},
  \code{\link{toleranceRange<-,msgfPar,numeric-method}},
  \code{\link{toleranceUnit}},
  \code{\link{toleranceUnit,msgfPar-method}},
  \code{\link{toleranceUnit<-}},
  \code{\link{toleranceUnit<-,msgfPar,character-method}}
}

