\name{createFileCall}
\alias{createFileCall}
\title{Create terminal compatible strings of input and output names}
\usage{
createFileCall(rawfile, savename)
}
\arguments{
  \item{rawfile}{The location of the spectrum file}

  \item{savename}{The location where the results should be
  saved}
}
\value{
A character vector that is compatible with the OS that can
be concatenated with the MS-GF+ call
}
\description{
Create terminal compatible strings of input and output
names
}

