% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{db}
\alias{db}
\alias{db,msgfPar-method}
\alias{db<-}
\alias{db<-,msgfPar,character-method}
\title{Get and set database in msgfPar objects}
\usage{
db(object)

db(object) <- value

\S4method{db}{msgfPar}(object)

\S4method{db}{msgfPar,character}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{A string matching the location of a fasta file}
}
\value{
In case of the getter a character vector with the location of the
database file
}
\description{
These functions allow you to retrieve and set the location of the database
fasta file in the msgfPar object of interest
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the database location

\item \code{object = msgfPar,value = character}: Set the database location
}}
\examples{
parameters <- msgfPar()
db(parameters) <- system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta')
db(parameters)
}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{chargeRange,msgfPar-method}},
  \code{\link{chargeRange<-}},
  \code{\link{chargeRange<-,msgfPar,msgfParChargeRange-method}},
  \code{\link{chargeRange<-,msgfPar,numeric-method}};
  \code{\link{enzyme}},
  \code{\link{enzyme,msgfPar-method}},
  \code{\link{enzyme<-}},
  \code{\link{enzyme<-,msgfPar,character-method}},
  \code{\link{enzyme<-,msgfPar,msgfParEnzyme-method}},
  \code{\link{enzyme<-,msgfPar,numeric-method}};
  \code{\link{fragmentation}},
  \code{\link{fragmentation,msgfPar-method}},
  \code{\link{fragmentation<-}},
  \code{\link{fragmentation<-,msgfPar,character-method}},
  \code{\link{fragmentation<-,msgfPar,msgfParFragmentation-method}},
  \code{\link{fragmentation<-,msgfPar,numeric-method}};
  \code{\link{instrument}},
  \code{\link{instrument,msgfPar-method}},
  \code{\link{instrument<-}},
  \code{\link{instrument<-,msgfPar,character-method}},
  \code{\link{instrument<-,msgfPar,msgfParInstrument-method}},
  \code{\link{instrument<-,msgfPar,numeric-method}};
  \code{\link{isotopeError}},
  \code{\link{isotopeError,msgfPar-method}},
  \code{\link{isotopeError<-}},
  \code{\link{isotopeError<-,msgfPar,msgfParIsotopeError-method}},
  \code{\link{isotopeError<-,msgfPar,numeric-method}};
  \code{\link{lengthRange}},
  \code{\link{lengthRange,msgfPar-method}},
  \code{\link{lengthRange<-}},
  \code{\link{lengthRange<-,msgfPar,msgfParLengthRange-method}},
  \code{\link{lengthRange<-,msgfPar,numeric-method}};
  \code{\link{matches}},
  \code{\link{matches,msgfPar-method}},
  \code{\link{matches<-}},
  \code{\link{matches<-,msgfPar,msgfParMatches-method}},
  \code{\link{matches<-,msgfPar,numeric-method}};
  \code{\link{mods}}, \code{\link{mods,msgfPar-method}},
  \code{\link{mods<-}},
  \code{\link{mods<-,msgfPar,msgfParModificationList-method}},
  \code{\link{nMod}}, \code{\link{nMod,msgfPar-method}},
  \code{\link{nMod<-}},
  \code{\link{nMod<-,msgfPar,numeric-method}};
  \code{\link{ntt}}, \code{\link{ntt,msgfPar-method}},
  \code{\link{ntt<-}},
  \code{\link{ntt<-,msgfPar,msgfParNtt-method}},
  \code{\link{ntt<-,msgfPar,numeric-method}};
  \code{\link{protocol}},
  \code{\link{protocol,msgfPar-method}},
  \code{\link{protocol<-}},
  \code{\link{protocol<-,msgfPar,character-method}},
  \code{\link{protocol<-,msgfPar,msgfParProtocol-method}},
  \code{\link{protocol<-,msgfPar,numeric-method}};
  \code{\link{tda}}, \code{\link{tda,msgfPar-method}},
  \code{\link{tda<-}},
  \code{\link{tda<-,msgfPar,logical-method}},
  \code{\link{tda<-,msgfPar,msgfParTda-method}};
  \code{\link{tolerance}},
  \code{\link{tolerance,msgfPar-method}},
  \code{\link{tolerance<-}},
  \code{\link{tolerance<-,msgfPar,character-method}},
  \code{\link{tolerance<-,msgfPar,msgfParTolerance-method}},
  \code{\link{toleranceRange}},
  \code{\link{toleranceRange,msgfPar-method}},
  \code{\link{toleranceRange<-}},
  \code{\link{toleranceRange<-,msgfPar,numeric-method}},
  \code{\link{toleranceUnit}},
  \code{\link{toleranceUnit,msgfPar-method}},
  \code{\link{toleranceUnit<-}},
  \code{\link{toleranceUnit<-,msgfPar,character-method}}
}

