% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{fragmentation}
\alias{fragmentation}
\alias{fragmentation,msgfPar-method}
\alias{fragmentation<-}
\alias{fragmentation<-,msgfPar,character-method}
\alias{fragmentation<-,msgfPar,msgfParFragmentation-method}
\alias{fragmentation<-,msgfPar,numeric-method}
\title{Get and set fragmentation in msgfPar objects}
\usage{
fragmentation(object)

fragmentation(object) <- value

\S4method{fragmentation}{msgfPar}(object)

\S4method{fragmentation}{msgfPar,numeric}(object) <- value

\S4method{fragmentation}{msgfPar,character}(object) <- value

\S4method{fragmentation}{msgfPar,msgfParFragmentation}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either an integer, string or msgfParFragmentation object}
}
\value{
In case of the getter a named integer
}
\description{
These functions allow you to retrieve and set the fragmentation method used
during acquisition
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the fragmentation method currently used

\item \code{object = msgfPar,value = numeric}: Set the fragmentation method using the key for the
method

\item \code{object = msgfPar,value = character}: Set the fragmentation method using the name of the
method

\item \code{object = msgfPar,value = msgfParFragmentation}: Set the fragmentation method using an
msgfParFragmentation object
}}
\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
fragmentation(parameters) <- 'CID'
fragmentation(parameters) <- 3
fragmentation(parameters)
}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{chargeRange,msgfPar-method}},
  \code{\link{chargeRange<-}},
  \code{\link{chargeRange<-,msgfPar,msgfParChargeRange-method}},
  \code{\link{chargeRange<-,msgfPar,numeric-method}};
  \code{\link{db}}, \code{\link{db,msgfPar-method}},
  \code{\link{db<-}},
  \code{\link{db<-,msgfPar,character-method}};
  \code{\link{enzyme}},
  \code{\link{enzyme,msgfPar-method}},
  \code{\link{enzyme<-}},
  \code{\link{enzyme<-,msgfPar,character-method}},
  \code{\link{enzyme<-,msgfPar,msgfParEnzyme-method}},
  \code{\link{enzyme<-,msgfPar,numeric-method}};
  \code{\link{instrument}},
  \code{\link{instrument,msgfPar-method}},
  \code{\link{instrument<-}},
  \code{\link{instrument<-,msgfPar,character-method}},
  \code{\link{instrument<-,msgfPar,msgfParInstrument-method}},
  \code{\link{instrument<-,msgfPar,numeric-method}};
  \code{\link{isotopeError}},
  \code{\link{isotopeError,msgfPar-method}},
  \code{\link{isotopeError<-}},
  \code{\link{isotopeError<-,msgfPar,msgfParIsotopeError-method}},
  \code{\link{isotopeError<-,msgfPar,numeric-method}};
  \code{\link{lengthRange}},
  \code{\link{lengthRange,msgfPar-method}},
  \code{\link{lengthRange<-}},
  \code{\link{lengthRange<-,msgfPar,msgfParLengthRange-method}},
  \code{\link{lengthRange<-,msgfPar,numeric-method}};
  \code{\link{matches}},
  \code{\link{matches,msgfPar-method}},
  \code{\link{matches<-}},
  \code{\link{matches<-,msgfPar,msgfParMatches-method}},
  \code{\link{matches<-,msgfPar,numeric-method}};
  \code{\link{mods}}, \code{\link{mods,msgfPar-method}},
  \code{\link{mods<-}},
  \code{\link{mods<-,msgfPar,msgfParModificationList-method}},
  \code{\link{nMod}}, \code{\link{nMod,msgfPar-method}},
  \code{\link{nMod<-}},
  \code{\link{nMod<-,msgfPar,numeric-method}};
  \code{\link{ntt}}, \code{\link{ntt,msgfPar-method}},
  \code{\link{ntt<-}},
  \code{\link{ntt<-,msgfPar,msgfParNtt-method}},
  \code{\link{ntt<-,msgfPar,numeric-method}};
  \code{\link{protocol}},
  \code{\link{protocol,msgfPar-method}},
  \code{\link{protocol<-}},
  \code{\link{protocol<-,msgfPar,character-method}},
  \code{\link{protocol<-,msgfPar,msgfParProtocol-method}},
  \code{\link{protocol<-,msgfPar,numeric-method}};
  \code{\link{tda}}, \code{\link{tda,msgfPar-method}},
  \code{\link{tda<-}},
  \code{\link{tda<-,msgfPar,logical-method}},
  \code{\link{tda<-,msgfPar,msgfParTda-method}};
  \code{\link{tolerance}},
  \code{\link{tolerance,msgfPar-method}},
  \code{\link{tolerance<-}},
  \code{\link{tolerance<-,msgfPar,character-method}},
  \code{\link{tolerance<-,msgfPar,msgfParTolerance-method}},
  \code{\link{toleranceRange}},
  \code{\link{toleranceRange,msgfPar-method}},
  \code{\link{toleranceRange<-}},
  \code{\link{toleranceRange<-,msgfPar,numeric-method}},
  \code{\link{toleranceUnit}},
  \code{\link{toleranceUnit,msgfPar-method}},
  \code{\link{toleranceUnit<-}},
  \code{\link{toleranceUnit<-,msgfPar,character-method}}
}

