% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getMSGFpar}
\alias{getMSGFpar}
\title{Get a string representation of an msgfPar-related object}
\usage{
getMSGFpar(object)
}
\arguments{
\item{object}{An msgfPar object or a related object}
}
\value{
A string that can be appended to a \code{system()} call to specify 
the parameters for the MSGF+ analysis
}
\description{
The string representation is defined as the arguments that should get
appended to the call when running MSGF+ in the terminal/command prompt
}
\examples{
parameters <- msgfPar(
                      database=system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'),
                      tolerance='20 ppm',
                      instrument='TOF',
                      enzyme='Lys-C'
                     )
getMSGFpar(parameters)

}
\seealso{
\code{\link{msgfPar-class}}
}
