% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{instrument}
\alias{instrument}
\alias{instrument<-}
\alias{instrument,msgfPar-method}
\alias{instrument<-,msgfPar,numeric-method}
\alias{instrument<-,msgfPar,character-method}
\alias{instrument<-,msgfPar,msgfParInstrument-method}
\title{Get and set instrument in msgfPar objects}
\usage{
instrument(object)

instrument(object) <- value

\S4method{instrument}{msgfPar}(object)

\S4method{instrument}{msgfPar,numeric}(object) <- value

\S4method{instrument}{msgfPar,character}(object) <- value

\S4method{instrument}{msgfPar,msgfParInstrument}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either an integer, string or msgfParInstrument object}
}
\value{
In case of the getter a named integer
}
\description{
These functions allow you to retrieve and set the instrument type used for
acquisition
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the instrument currently used

\item \code{object = msgfPar,value = numeric}: Set the instrument using the key for the instrument

\item \code{object = msgfPar,value = character}: Set the instrument using the name of the instrument

\item \code{object = msgfPar,value = msgfParInstrument}: Set the instrument using an msgfParInstrument object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
instrument(parameters) <- 'TOF'
instrument(parameters) <- 3
instrument(parameters)

}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{enzyme}},
  \code{\link{fragmentation}}, \code{\link{isotopeError}},
  \code{\link{lengthRange}}, \code{\link{matches}},
  \code{\link{mods}}, \code{\link{ntt}},
  \code{\link{protocol}}, \code{\link{tda}},
  \code{\link{tolerance}}
}
