% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{isotopeError}
\alias{isotopeError}
\alias{isotopeError<-}
\alias{isotopeError,msgfPar-method}
\alias{isotopeError<-,msgfPar,numeric-method}
\alias{isotopeError<-,msgfPar,msgfParIsotopeError-method}
\title{Get and set isotope error in msgfPar objects}
\usage{
isotopeError(object)

isotopeError(object) <- value

\S4method{isotopeError}{msgfPar}(object)

\S4method{isotopeError}{msgfPar,numeric}(object) <- value

\S4method{isotopeError}{msgfPar,msgfParIsotopeError}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either an integer vector or an msgfParIsotopeError object}
}
\value{
In case of the getter an integer vector
}
\description{
These functions allow you to retrieve and set the isotope error used when 
calculating the parent ion error range
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the isotope error currently used

\item \code{object = msgfPar,value = numeric}: Set the isotope error with an integer vector

\item \code{object = msgfPar,value = msgfParIsotopeError}: Set the isotope error with an msgfParIsotopeError 
object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
isotopeError(parameters) <- c(0, 3)
isotopeError(parameters)
}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{enzyme}},
  \code{\link{fragmentation}}, \code{\link{instrument}},
  \code{\link{lengthRange}}, \code{\link{matches}},
  \code{\link{mods}}, \code{\link{ntt}},
  \code{\link{protocol}}, \code{\link{tda}},
  \code{\link{tolerance}}
}
