\docType{methods}
\name{length,msgfPar-method}
\alias{length,msgfPar-method}
\title{Report the length of an msgfPar object}
\usage{
\S4method{length}{msgfPar}(x)
}
\arguments{
  \item{x}{An msgfPar object}
}
\value{
A \code{numeric} giving the lengthof the object
}
\description{
The length of an msgfPar object is defined as 1 if a
database is present and 0 otherwise
}
\seealso{
\code{\link{msgfPar-class}}
}

