% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{lengthRange}
\alias{lengthRange}
\alias{lengthRange<-}
\alias{lengthRange,msgfPar-method}
\alias{lengthRange<-,msgfPar,numeric-method}
\alias{lengthRange<-,msgfPar,msgfParLengthRange-method}
\title{Get and set peptide length range in msgfPar objects}
\usage{
lengthRange(object)

lengthRange(object) <- value

\S4method{lengthRange}{msgfPar}(object)

\S4method{lengthRange}{msgfPar,numeric}(object) <- value

\S4method{lengthRange}{msgfPar,msgfParLengthRange}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either an integer vector or an msgfParLengthRange object}
}
\value{
In case of the getter an integer vector of length 2 giving the lower
and upper bounds of the length range
}
\description{
These functions allow you to retrieve and set the residue length allowed for
the peptides search for in MS-GF+
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the lower and upper bounds of peptide lengths

\item \code{object = msgfPar,value = numeric}: Set the lower and upper bounds of peptide lengths 
using an integer vector

\item \code{object = msgfPar,value = msgfParLengthRange}: Set the lower and upper bounds of peptide lengths 
using an msgfParLengthRange
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
lengthRange(parameters) <- c(6, 40)
lengthRange(parameters)

}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{enzyme}},
  \code{\link{fragmentation}}, \code{\link{instrument}},
  \code{\link{isotopeError}}, \code{\link{matches}},
  \code{\link{mods}}, \code{\link{ntt}},
  \code{\link{protocol}}, \code{\link{tda}},
  \code{\link{tolerance}}
}
