% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{msgfPar-class}
\alias{msgfPar-class}
\title{A class to contain parameters used in an MS-GF+ analysis}
\description{
This class collects and stores parameters for an MS-GF+ analysis and is the
starting point for peptide identification
}
\details{
This class contains a range of other classes, each handling a different set
of parameters. Often these classes are simple containers that only takes care
of errorchecking and generating command line arguments, but in some cases, as
with msgfParModificationList, the class is a bit more complex.
}
\section{Objects from the class}{

Objects can be created using the \code{\link{msgfPar}} constructor, or with \code{\link{msgfParGUI}} for a simple graphical user interface
}

\section{Slots}{

\describe{
 \item{\code{database}:}{The location of the database fasta file used for the analysis.}
 \item{\code{tolerance}:}{An \code{msgfParTolerance} object holding the m/z tolerance used in the search.}
 \item{\code{isotopeError}:}{An \code{msgfParIsotopeError} object holding the isotope errors permitted in the search.}
 \item{\code{tda}:}{An \code{msgfParTda} object saying whether FDR should be estimated using the target-decoy approach.}
 \item{\code{fragmentation}:}{An \code{msgfParFragmentation} object holding the type of fragmentation expected from the experiment.}
 \item{\code{instrument}:}{An \code{msgfParInstrument} object holding which type of instrument was used for collecting the data.}
 \item{\code{enzyme}:}{An \code{msgfParEnzyme} object holding which enzyme was used for digestion}
 \item{\code{protocol}:}{An \code{msgfParProtocol} object defining whether a specific protocol should be used in the search.}
 \item{\code{ntt}:}{An \code{msgfParNtt} object defining the number of tolerable termini allowed in the peptides.}
 \item{\code{modification}:}{An \code{msgfParModificationList} object holding the modifications accepted in the search.}
 \item{\code{lengthRange}:}{An \code{msgfParLengthRange} object setting the limits on the peptide length in residues that the search allows.}
 \item{\code{chargeRange}:}{An \code{msgfParChargeRange} object defining which charges should be included in the search.}
 \item{\code{matches}:}{An \code{msgfParMatches} object defining the number of matches per PSM that gets reported in the output.}
}
}

\section{Methods}{

\describe{
 \item{\code{\link{getMSGFpar}}:}{Get a string representation of the object that can be directly used in a system call.}
 \item{\code{\link{runMSGF}}:}{Start an MS-GF+ search based on the msgfPar object.}
 \item{\code{\link{database}}:}{Retrieves the database location from the msgfPar object.}
 \item{\code{\link{database<-}}:}{Sets the database location for the msgfPar object.}
}
}
\references{
\url{http://proteomics.ucsd.edu/Software/MSGFPlus.html}
}
\seealso{
\code{\link{msgfParGUI}}

Other msgfParClasses: \code{\link{msgfParChargeRange}},
  \code{\link{msgfParChargeRange-class}};
  \code{\link{msgfParEnzyme}}; \code{\link{msgfPar}}
}

