% Generated by roxygen2 (4.0.2): do not edit by hand
\name{msgfPar}
\alias{msgfPar}
\title{Constructor for the msgfPar class}
\usage{
msgfPar(database, tolerance, isotopeError, tda, fragmentation, instrument,
  enzyme, protocol, ntt, modification, lengthRange, chargeRange, matches)
}
\arguments{
\item{database}{The location of the fasta file to use as search database}

\item{tolerance}{The parent ion tolerance to use. In simple cases a string in
the form '20 ppm' or '1 Da' or an msgfParTolerance object if asymmetric
tolerance is desired}

\item{isotopeError}{The range of isotope errors used to correct for
non-monoisotopic peaks. Either a numeric vector of length 2 specifying the
lower and upper bounds of the range, or an msgfParIsotopeError object}

\item{tda}{\code{Logical} Should Target-Decoy approach be used to calculate
FDR values.}

\item{fragmentation}{An integer specifying which fragmentation has been used
during data acquisition. See details.}

\item{instrument}{An integer specifying the type of instrument used during
data acquisition. See details.}

\item{enzyme}{An integer or name specifying the enzyme that has been used
for protein digestion. See details.}

\item{protocol}{An integer or name specifying the type of preparation that
has been done for the samples. See details.}

\item{ntt}{An integer specifying the cleavage specificity (Number of
Tolerable Termini). 2 only allows fully tryptic peptide (if trypsin is used),
1 allows semitryptic peptides and 0 allows unspecific peptides}

\item{modification}{An msgfParModificationList object or a list containing
the named elements nMod and modifications containing respectively an integer
with the number of allowed modifications per petide and the modifications to
search for as msgfParModification}

\item{lengthRange}{A two element vector containing the lower and upper bounds
of the residue length to search for}

\item{chargeRange}{A two element vector containing the lower and upper bounds
of the charge range to search for}

\item{matches}{The number of matches to report per spectrum}
}
\value{
An msgfPar object
}
\description{
This function creates an msgfPar object with the specified parameters. If
some parameters have not been specified they will not be part of the MS-GF+
call and MS-GF+'s own defaults kicks in; Consult the MS-GF+ documentation for
these. Note however that at least a database file is required to run an
analysis.
}
\details{
Please consult the MS-GF+ documentation for full description of the
parameters

Fragmentation is usually specified as an integer according to the following
lookup
\describe{
  \item{0}{As written in the spectrum or CID if no info}
  \item{1}{CID}
  \item{2}{ETD}
  \item{3}{HCD}
  \item{4}{Merge spectra from the same precursor}
}
It is possible to use the full name of the description for a more litteral
function call

Instrument can likewise be specified as an integer or as a name according to
this list
\describe{
  \item{0}{LowRes}
  \item{1}{HighRes}
  \item{2}{TOF}
  \item{3}{QExactive}
}

Enymes are specified in the same manner using the following list
\describe{
  \item{0}{Unspecific cleavage}
  \item{1}{Trypsin}
  \item{2}{Chymotrypsin}
  \item{3}{Lys-C}
  \item{4}{Lys-N}
  \item{5}{glutamyl endopeptidase (Glu-C)}
  \item{6}{Arg-C}
  \item{7}{Asp-N}
  \item{8}{alphaLP}
  \item{9}{No cleavage}
}

The protocol informs MS-GF+ whether a special sample treatment has been
performed as part of the analysis. The protocol is specified according to
the following list
\describe{
  \item{0}{No protocol}
  \item{1}{Phosphorylation}
  \item{2}{iTRAQ}
  \item{3}{iTRAQPhospho}
}
}
\examples{
# Example of specifying all parameters - usually not necessary
parameters <- msgfPar(
                      database=system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'),
                      tolerance='20 ppm',
                      isotopeError=c(0, 2),
                      tda=TRUE,
                      fragmentation='CID',
                      instrument='TOF',
                      enzyme='Lys-C',
                      protocol='No protocol',
                      ntt=2,
                      modification=list(
                          nMod=2,
                          modifications=list(
                              list(name='Carbamidomethyl',
                                   composition='C2H3N1O1',
                                   residues='C',
                                   type='fix',
                                   position='any'
                              ),
                              list(name='Oxidation',
                                   mass=15.994915,
                                   residues='M',
                                   type='opt',
                                   position='any'
                              )
                          )
                      ),
                      lengthRange=c(6,40),
                      chargeRange=c(2,7),
                      matches=1
                     )
parameters
}
\references{
\href{http://proteomics.ucsd.edu/Software/MSGFPlus.html}{MS-GF+}
}
\seealso{
Other msgfParClasses: \code{\link{[[,msgfParModificationList,numeric,missing-method}},
  \code{\link{[[<-,msgfParModificationList,numeric,missing,msgfParModification-method}},
  \code{\link{getMSGFpar,msgfParModificationList-method}},
  \code{\link{length,msgfParModificationList-method}},
  \code{\link{msgfParModificationList}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{show,msgfParModificationList-method}};
  \code{\link{getMSGFpar,msgfPar-method}},
  \code{\link{length,msgfPar-method}},
  \code{\link{msgfPar-class}},
  \code{\link{runMSGF,msgfPar-method}},
  \code{\link{show,msgfPar-method}};
  \code{\link{getMSGFpar,msgfParChargeRange-method}},
  \code{\link{length,msgfParChargeRange-method}},
  \code{\link{msgfParChargeRange}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{show,msgfParChargeRange-method}};
  \code{\link{getMSGFpar,msgfParEnzyme-method}},
  \code{\link{length,msgfParEnzyme-method}},
  \code{\link{msgfParEnzyme}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{show,msgfParEnzyme-method}};
  \code{\link{getMSGFpar,msgfParFragmentation-method}},
  \code{\link{length,msgfParFragmentation-method}},
  \code{\link{msgfParFragmentation}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{show,msgfParFragmentation-method}};
  \code{\link{getMSGFpar,msgfParInstrument-method}},
  \code{\link{length,msgfParInstrument-method}},
  \code{\link{msgfParInstrument}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{show,msgfParInstrument-method}};
  \code{\link{getMSGFpar,msgfParIsotopeError-method}},
  \code{\link{length,msgfParIsotopeError-method}},
  \code{\link{msgfParIsotopeError}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{show,msgfParIsotopeError-method}};
  \code{\link{getMSGFpar,msgfParLengthRange-method}},
  \code{\link{length,msgfParLengthRange-method}},
  \code{\link{msgfParLengthRange}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{show,msgfParLengthRange-method}};
  \code{\link{getMSGFpar,msgfParMatches-method}},
  \code{\link{length,msgfParMatches-method}},
  \code{\link{msgfParMatches}},
  \code{\link{msgfParMatches-class}},
  \code{\link{show,msgfParMatches-method}};
  \code{\link{getMSGFpar,msgfParModification-method}},
  \code{\link{msgfParModification}},
  \code{\link{msgfParModification-class}},
  \code{\link{show,msgfParModification-method}};
  \code{\link{getMSGFpar,msgfParNtt-method}},
  \code{\link{length,msgfParNtt-method}},
  \code{\link{msgfParNtt}}, \code{\link{msgfParNtt-class}},
  \code{\link{show,msgfParNtt-method}};
  \code{\link{getMSGFpar,msgfParProtocol-method}},
  \code{\link{length,msgfParProtocol-method}},
  \code{\link{msgfParProtocol}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{show,msgfParProtocol-method}};
  \code{\link{getMSGFpar,msgfParTda-method}},
  \code{\link{length,msgfParTda-method}},
  \code{\link{msgfParTda}}, \code{\link{msgfParTda-class}},
  \code{\link{show,msgfParTda-method}};
  \code{\link{getMSGFpar,msgfParTolerance-method}},
  \code{\link{length,msgfParTolerance-method}},
  \code{\link{msgfParTolerance}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{show,msgfParTolerance-method}}
}

