% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{msgfPar getters amd setters}
\alias{chargeRange}
\alias{chargeRange,msgfPar-method}
\alias{chargeRange<-}
\alias{database}
\alias{database,msgfPar-method}
\alias{database<-}
\alias{enzyme}
\alias{enzyme,msgfPar-method}
\alias{enzyme<-}
\alias{fragmentation}
\alias{fragmentation,msgfPar-method}
\alias{fragmentation<-}
\alias{instrument}
\alias{instrument,msgfPar-method}
\alias{instrument<-}
\alias{isotopeError}
\alias{isotopeError,msgfPar-method}
\alias{isotopeError<-}
\alias{lengthRange}
\alias{lengthRange,msgfPar-method}
\alias{lengthRange<-}
\alias{matches}
\alias{matches,msgfPar-method}
\alias{matches<-}
\alias{modifications}
\alias{modifications,msgfPar-method}
\alias{modifications<-}
\alias{msgfPar getters amd setters}
\alias{nMod}
\alias{nMod,msgfPar-method}
\alias{nMod<-}
\alias{ntt}
\alias{ntt,msgfPar-method}
\alias{ntt<-}
\alias{protocol}
\alias{protocol,msgfPar-method}
\alias{protocol<-}
\alias{tda}
\alias{tda,msgfPar-method}
\alias{tda<-}
\alias{tolerance}
\alias{tolerance,msgfPar-method}
\alias{tolerance<-}
\alias{toleranceRange}
\alias{toleranceRange,msgfPar-method}
\alias{toleranceRange<-}
\alias{toleranceUnit}
\alias{toleranceUnit,msgfPar-method}
\alias{toleranceUnit<-}
\title{Get and set the location of the database file as specified in the msgfPar object}
\usage{
database(object)

database(object) <- value

chargeRange(object)

chargeRange(object) <- value

enzyme(object)

enzyme(object) <- value

fragmentation(object)

fragmentation(object) <- value

instrument(object)

instrument(object) <- value

isotopeError(object)

isotopeError(object) <- value

lengthRange(object)

lengthRange(object) <- value

matches(object)

matches(object) <- value

ntt(object)

ntt(object) <- value

protocol(object)

protocol(object) <- value

tda(object)

tda(object) <- value

tolerance(object)

tolerance(object) <- value

toleranceRange(object)

toleranceRange(object) <- value

toleranceUnit(object)

toleranceUnit(object) <- value

modifications(object)

modifications(object) <- value

nMod(object)

nMod(object) <- value

\S4method{database}{msgfPar}(object)

\S4method{chargeRange}{msgfPar}(object)

\S4method{enzyme}{msgfPar}(object)

\S4method{fragmentation}{msgfPar}(object)

\S4method{instrument}{msgfPar}(object)

\S4method{isotopeError}{msgfPar}(object)

\S4method{lengthRange}{msgfPar}(object)

\S4method{matches}{msgfPar}(object)

\S4method{ntt}{msgfPar}(object)

\S4method{protocol}{msgfPar}(object)

\S4method{tda}{msgfPar}(object)

\S4method{tolerance}{msgfPar}(object)

\S4method{toleranceRange}{msgfPar}(object)

\S4method{toleranceUnit}{msgfPar}(object)

\S4method{modifications}{msgfPar}(object)

\S4method{nMod}{msgfPar}(object)
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Depending on the method. See details}
}
\value{
Depending on the method. See details
}
\description{
This function extracts or replaces the string holding the database location
in the msgfPar object
}
\seealso{
\code{\link{msgfPar-class}}
}

