\docType{class}
\name{msgfParChargeRange-class}
\alias{msgfParChargeRange-class}
\title{A class handling charge ranges}
\description{
This class defines a charge range and provides methods to
get correct system call parameters.
}
\section{Slots}{
  \describe{ \item{\code{value}:}{A numeric vector of
  length 2 describing the upper and lower bounds of the
  charge range} }
}

