% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{msgfParEnzyme}
\alias{msgfParEnzyme}
\title{A class handling enzyme selection}
\description{
This class defines a digestion enzyme selection and provides methods to get correct system
call parameters.
}
\section{Slots}{

\describe{
 \item{\code{enzyme}:}{An integer specifiying the selection of enzyme. See the detail section of \code{\link{msgfPar}}}
}
}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}};
  \code{\link{msgfParChargeRange}},
  \code{\link{msgfParChargeRange-class}};
  \code{\link{msgfPar}}
}

