% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msgfParInstrument.R
\docType{class}
\name{msgfParInstrument-class}
\alias{getMSGFpar,msgfParInstrument-method}
\alias{length,msgfParInstrument-method}
\alias{msgfParInstrument}
\alias{msgfParInstrument-class}
\alias{show,msgfParInstrument-method}
\title{A class handling instrument types}
\usage{
\S4method{show}{msgfParInstrument}(object)

\S4method{length}{msgfParInstrument}(x)

\S4method{getMSGFpar}{msgfParInstrument}(object)

msgfParInstrument(instrument)
}
\arguments{
\item{object}{An msgfParInstrument object}

\item{x}{An msgfParInstrument object}

\item{instrument}{An integer specifying the instrument type}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParInstrument() An msgfParInstrument object.
}
\description{
This class defines an instrument type and provides methods to get correct
system
call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParInstrument object

\item \code{length}: Report the length of an msgfParInstrument
object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant
function call
}}
\section{Slots}{

\describe{
\item{\code{instrument}}{An integer specifying the instrument type}
}}
\examples{
instrument <- msgfParInstrument(1)
instrument <- msgfParInstrument('HighRes')
}
\seealso{
Other msgfParClasses: \code{\link{[[,msgfParModificationList,numeric,missing-method}},
  \code{\link{[[<-,msgfParModificationList,numeric,missing,msgfParModification-method}},
  \code{\link{getMSGFpar,msgfParModificationList-method}},
  \code{\link{length,msgfParModificationList-method}},
  \code{\link{msgfParModificationList}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{show,msgfParModificationList-method}};
  \code{\link{getMSGFpar,msgfPar-method}},
  \code{\link{length,msgfPar-method}},
  \code{\link{msgfPar-class}},
  \code{\link{runMSGF,msgfPar-method}},
  \code{\link{show,msgfPar-method}};
  \code{\link{getMSGFpar,msgfParChargeRange-method}},
  \code{\link{length,msgfParChargeRange-method}},
  \code{\link{msgfParChargeRange}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{show,msgfParChargeRange-method}};
  \code{\link{getMSGFpar,msgfParEnzyme-method}},
  \code{\link{length,msgfParEnzyme-method}},
  \code{\link{msgfParEnzyme}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{show,msgfParEnzyme-method}};
  \code{\link{getMSGFpar,msgfParFragmentation-method}},
  \code{\link{length,msgfParFragmentation-method}},
  \code{\link{msgfParFragmentation}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{show,msgfParFragmentation-method}};
  \code{\link{getMSGFpar,msgfParIsotopeError-method}},
  \code{\link{length,msgfParIsotopeError-method}},
  \code{\link{msgfParIsotopeError}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{show,msgfParIsotopeError-method}};
  \code{\link{getMSGFpar,msgfParLengthRange-method}},
  \code{\link{length,msgfParLengthRange-method}},
  \code{\link{msgfParLengthRange}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{show,msgfParLengthRange-method}};
  \code{\link{getMSGFpar,msgfParMatches-method}},
  \code{\link{length,msgfParMatches-method}},
  \code{\link{msgfParMatches}},
  \code{\link{msgfParMatches-class}},
  \code{\link{show,msgfParMatches-method}};
  \code{\link{getMSGFpar,msgfParModification-method}},
  \code{\link{msgfParModification}},
  \code{\link{msgfParModification-class}},
  \code{\link{show,msgfParModification-method}};
  \code{\link{getMSGFpar,msgfParNtt-method}},
  \code{\link{length,msgfParNtt-method}},
  \code{\link{msgfParNtt}}, \code{\link{msgfParNtt-class}},
  \code{\link{show,msgfParNtt-method}};
  \code{\link{getMSGFpar,msgfParProtocol-method}},
  \code{\link{length,msgfParProtocol-method}},
  \code{\link{msgfParProtocol}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{show,msgfParProtocol-method}};
  \code{\link{getMSGFpar,msgfParTda-method}},
  \code{\link{length,msgfParTda-method}},
  \code{\link{msgfParTda}}, \code{\link{msgfParTda-class}},
  \code{\link{show,msgfParTda-method}};
  \code{\link{getMSGFpar,msgfParTolerance-method}},
  \code{\link{length,msgfParTolerance-method}},
  \code{\link{msgfParTolerance}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{show,msgfParTolerance-method}};
  \code{\link{msgfPar}}
}

