% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParIsotopeError.R
\docType{class}
\name{msgfParIsotopeError-class}
\alias{msgfParIsotopeError-class}
\alias{show,msgfParIsotopeError-method}
\alias{length,msgfParIsotopeError-method}
\alias{getMSGFpar,msgfParIsotopeError-method}
\alias{msgfParIsotopeError}
\title{A class handling isotope errors}
\usage{
\S4method{show}{msgfParIsotopeError}(object)

\S4method{length}{msgfParIsotopeError}(x)

\S4method{getMSGFpar}{msgfParIsotopeError}(object)

msgfParIsotopeError(range)
}
\arguments{
\item{object}{An msgfParIsotopeError object}

\item{x}{An msgfParIsotopeError object}

\item{range}{An integer vector with isotopes}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParIsotopeError() An msgfParIsotopeError object.
}
\description{
This class defines a set of isotopes that should be included for error 
correction and provides methods to get correct system call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParIsotopeError object

\item \code{length}: Report the length of an msgfParIsotopeError 
object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant 
function call
}}

\section{Slots}{

\describe{
\item{\code{range}}{An integer vector with lower and upper bounds of isotopes to 
error correct}
}}

\examples{
isotopeError <- msgfParIsotopeError(c(0, 2))

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
