% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{msgfParModificationList-class}
\alias{[[,msgfParModificationList,numeric,missing-method}
\alias{[[<-,msgfParModificationList,numeric,missing,msgfParModification-method}
\alias{getMSGFpar,msgfParModificationList-method}
\alias{length,msgfParModificationList-method}
\alias{msgfParModificationList}
\alias{msgfParModificationList-class}
\alias{show,msgfParModificationList-method}
\title{A class handling a list of modifications}
\usage{
\S4method{show}{msgfParModificationList}(object)

\S4method{length}{msgfParModificationList}(x)

\S4method{getMSGFpar}{msgfParModificationList}(object)

\S4method{[[}{msgfParModificationList,numeric,missing}(x, i, j, ...)

\S4method{[[}{msgfParModificationList,numeric,missing,msgfParModification}(x, i,
  j, ...) <- value

msgfParModificationList(nMod, modifications = list())
}
\arguments{
\item{object}{An msgfParModificationList object}

\item{x}{An msgfParModificationList object}

\item{i}{The index of the modification}

\item{j}{Ignored}

\item{...}{Ignored}

\item{value}{An msgfParModification object}

\item{nMod}{The maximum allowed number of modifications to expect on any peptide}

\item{modifications}{A list of \code{\linkS4class{msgfParModification}} objects}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For '[[' A msgfParModification object

For msgfParModificationList() An msgfParModificationList object.
}
\description{
This class defines a set of modifications and a maximum number of
modifications allowed per peptide.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParModificationList object

\item \code{length}: Report the length of an msgfParModificationList object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant function call

\item \code{[[}: Get the i'th modification

\item \code{[[<-}: Set or change the i'th modification
}}
\section{Slots}{

\describe{
\item{\code{nMod}}{The maximum allowed number of modifications to expect on any peptide}

\item{\code{modifications}}{A list of \code{\linkS4class{msgfParModification}} objects}
}}
\examples{
modification1 <- msgfParModification(
                                     name='Carbamidomethyl',
                                     composition='C2H3N1O1',
                                     residues='C',
                                     type='fix',
                                     position='any'
                                    )
modification2 <- msgfParModification(
                                     name='Oxidation',
                                     mass=15.994915,
                                     residues='M',
                                     type='opt',
                                     position='any'
                                    )
modificationlist <- msgfParModificationList(
                                            nMod=2,
                                            modifications=list(
                                                modification1,
                                                modification2
                                            )
                                           )
modificationlist[[3]] <- msgfParModification(
                                             name='Gln->pyro-Glu',
                                             composition='H-3N-1',
                                             residues='Q',
                                             type='opt',
                                             position='N-term'
                                            )
}
\seealso{
Other msgfParClasses: \code{\link{getMSGFpar,msgfPar-method}},
  \code{\link{length,msgfPar-method}},
  \code{\link{msgfPar-class}},
  \code{\link{runMSGF,msgfPar-method}},
  \code{\link{show,msgfPar-method}};
  \code{\link{getMSGFpar,msgfParChargeRange-method}},
  \code{\link{length,msgfParChargeRange-method}},
  \code{\link{msgfParChargeRange}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{show,msgfParChargeRange-method}};
  \code{\link{getMSGFpar,msgfParEnzyme-method}},
  \code{\link{length,msgfParEnzyme-method}},
  \code{\link{msgfParEnzyme}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{show,msgfParEnzyme-method}};
  \code{\link{getMSGFpar,msgfParFragmentation-method}},
  \code{\link{length,msgfParFragmentation-method}},
  \code{\link{msgfParFragmentation}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{show,msgfParFragmentation-method}};
  \code{\link{getMSGFpar,msgfParInstrument-method}},
  \code{\link{length,msgfParInstrument-method}},
  \code{\link{msgfParInstrument}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{show,msgfParInstrument-method}};
  \code{\link{getMSGFpar,msgfParIsotopeError-method}},
  \code{\link{length,msgfParIsotopeError-method}},
  \code{\link{msgfParIsotopeError}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{show,msgfParIsotopeError-method}};
  \code{\link{getMSGFpar,msgfParLengthRange-method}},
  \code{\link{length,msgfParLengthRange-method}},
  \code{\link{msgfParLengthRange}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{show,msgfParLengthRange-method}};
  \code{\link{getMSGFpar,msgfParMatches-method}},
  \code{\link{length,msgfParMatches-method}},
  \code{\link{msgfParMatches}},
  \code{\link{msgfParMatches-class}},
  \code{\link{show,msgfParMatches-method}};
  \code{\link{getMSGFpar,msgfParModification-method}},
  \code{\link{msgfParModification}},
  \code{\link{msgfParModification-class}},
  \code{\link{show,msgfParModification-method}};
  \code{\link{getMSGFpar,msgfParNtt-method}},
  \code{\link{length,msgfParNtt-method}},
  \code{\link{msgfParNtt}}, \code{\link{msgfParNtt-class}},
  \code{\link{show,msgfParNtt-method}};
  \code{\link{getMSGFpar,msgfParProtocol-method}},
  \code{\link{length,msgfParProtocol-method}},
  \code{\link{msgfParProtocol}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{show,msgfParProtocol-method}};
  \code{\link{getMSGFpar,msgfParTda-method}},
  \code{\link{length,msgfParTda-method}},
  \code{\link{msgfParTda}}, \code{\link{msgfParTda-class}},
  \code{\link{show,msgfParTda-method}};
  \code{\link{getMSGFpar,msgfParTolerance-method}},
  \code{\link{length,msgfParTolerance-method}},
  \code{\link{msgfParTolerance}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{show,msgfParTolerance-method}};
  \code{\link{msgfPar}}
}

