% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{MSGFplus-package}
\alias{MSGFplus-package}
\title{An interface between R and MS-GF+}
\description{
This package comes bundled with the peptide database search tool MS-GF+ and
provides means of running peptide searches from within R with automatic
result import and easy handling of parameters.
}
\details{
The version bundled with this package is: MS-GF+ Beta (v9881) (11/11/2013)
}
\author{
Thomas Dybdal Pedersen with contributions from Laurent Gatto
}

