\docType{methods}
\name{show,msgfPar-method}
\alias{show,msgfPar-method}
\title{Show method for msgfPar objects}
\usage{
\S4method{show}{msgfPar}(object)
}
\arguments{
  \item{object}{An msgfPar object}
}
\value{
A description of the content of the msgfPar object
}
\description{
This method report a general summary of the parameters
specified in the msgfPar object
}
\seealso{
\code{\link{msgfPar-class}}
}

