% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{tolerance}
\alias{tolerance}
\alias{tolerance,msgfPar-method}
\alias{tolerance<-}
\alias{tolerance<-,msgfPar,character-method}
\alias{tolerance<-,msgfPar,msgfParTolerance-method}
\alias{toleranceRange}
\alias{toleranceRange,msgfPar-method}
\alias{toleranceRange<-}
\alias{toleranceRange<-,msgfPar,numeric-method}
\alias{toleranceUnit}
\alias{toleranceUnit,msgfPar-method}
\alias{toleranceUnit<-}
\alias{toleranceUnit<-,msgfPar,character-method}
\title{Get and set the parent tolerance in msgfPar objects}
\usage{
tolerance(object)

tolerance(object) <- value

toleranceRange(object)

toleranceRange(object) <- value

toleranceUnit(object)

toleranceUnit(object) <- value

\S4method{tolerance}{msgfPar}(object)

\S4method{toleranceRange}{msgfPar}(object)

\S4method{toleranceUnit}{msgfPar}(object)

\S4method{toleranceRange}{msgfPar,numeric}(object) <- value

\S4method{toleranceUnit}{msgfPar,character}(object) <- value

\S4method{tolerance}{msgfPar,character}(object) <- value

\S4method{tolerance}{msgfPar,msgfParTolerance}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{For tolerance a character vector of length 2, each element of
the form '<value> <unit>'. For toleranceUnit a string. For toleranceRange a
numeric vector of length 2.}
}
\value{
For tolerance a character vector with the lower and upper tolerance
limit with unit. For toleranceUnit a string with the unit used. For
toleranceRange a numeric vector with lower and upper tolerance limit.
}
\description{
These functions allow you to retrieve and set the tolerance used for matching
parent ions to peptides in the database
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the lower and upper bounds of the tolerance

\item \code{msgfPar}: Get the lower and upper bounds of the tolerance

\item \code{msgfPar}: Get the unit the tolerance is measured in

\item \code{object = msgfPar,value = numeric}: Set the lower and upper bounds of the tolerance using a
numeric vector of length 2

\item \code{object = msgfPar,value = character}: Set the unit the tolerance is meassured in

\item \code{object = msgfPar,value = character}: Set the lower and upper bounds of the tolerance using a
character vector of length 2, where each element is of the form '<value>
<unit>'

\item \code{object = msgfPar,value = msgfParTolerance}: Set the lower and upper bounds of the tolerance using
an msgfParTolerance object
}}
\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
tolerance(parameters) <- c('20 ppm', '20 ppm')
toleranceUnit(parameters) <- 'Da'
toleranceRange(parameters) <- c(1.5, 1.5)
tolerance(parameters)
}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{chargeRange,msgfPar-method}},
  \code{\link{chargeRange<-}},
  \code{\link{chargeRange<-,msgfPar,msgfParChargeRange-method}},
  \code{\link{chargeRange<-,msgfPar,numeric-method}};
  \code{\link{db}}, \code{\link{db,msgfPar-method}},
  \code{\link{db<-}},
  \code{\link{db<-,msgfPar,character-method}};
  \code{\link{enzyme}},
  \code{\link{enzyme,msgfPar-method}},
  \code{\link{enzyme<-}},
  \code{\link{enzyme<-,msgfPar,character-method}},
  \code{\link{enzyme<-,msgfPar,msgfParEnzyme-method}},
  \code{\link{enzyme<-,msgfPar,numeric-method}};
  \code{\link{fragmentation}},
  \code{\link{fragmentation,msgfPar-method}},
  \code{\link{fragmentation<-}},
  \code{\link{fragmentation<-,msgfPar,character-method}},
  \code{\link{fragmentation<-,msgfPar,msgfParFragmentation-method}},
  \code{\link{fragmentation<-,msgfPar,numeric-method}};
  \code{\link{instrument}},
  \code{\link{instrument,msgfPar-method}},
  \code{\link{instrument<-}},
  \code{\link{instrument<-,msgfPar,character-method}},
  \code{\link{instrument<-,msgfPar,msgfParInstrument-method}},
  \code{\link{instrument<-,msgfPar,numeric-method}};
  \code{\link{isotopeError}},
  \code{\link{isotopeError,msgfPar-method}},
  \code{\link{isotopeError<-}},
  \code{\link{isotopeError<-,msgfPar,msgfParIsotopeError-method}},
  \code{\link{isotopeError<-,msgfPar,numeric-method}};
  \code{\link{lengthRange}},
  \code{\link{lengthRange,msgfPar-method}},
  \code{\link{lengthRange<-}},
  \code{\link{lengthRange<-,msgfPar,msgfParLengthRange-method}},
  \code{\link{lengthRange<-,msgfPar,numeric-method}};
  \code{\link{matches}},
  \code{\link{matches,msgfPar-method}},
  \code{\link{matches<-}},
  \code{\link{matches<-,msgfPar,msgfParMatches-method}},
  \code{\link{matches<-,msgfPar,numeric-method}};
  \code{\link{mods}}, \code{\link{mods,msgfPar-method}},
  \code{\link{mods<-}},
  \code{\link{mods<-,msgfPar,msgfParModificationList-method}},
  \code{\link{nMod}}, \code{\link{nMod,msgfPar-method}},
  \code{\link{nMod<-}},
  \code{\link{nMod<-,msgfPar,numeric-method}};
  \code{\link{ntt}}, \code{\link{ntt,msgfPar-method}},
  \code{\link{ntt<-}},
  \code{\link{ntt<-,msgfPar,msgfParNtt-method}},
  \code{\link{ntt<-,msgfPar,numeric-method}};
  \code{\link{protocol}},
  \code{\link{protocol,msgfPar-method}},
  \code{\link{protocol<-}},
  \code{\link{protocol<-,msgfPar,character-method}},
  \code{\link{protocol<-,msgfPar,msgfParProtocol-method}},
  \code{\link{protocol<-,msgfPar,numeric-method}};
  \code{\link{tda}}, \code{\link{tda,msgfPar-method}},
  \code{\link{tda<-}},
  \code{\link{tda<-,msgfPar,logical-method}},
  \code{\link{tda<-,msgfPar,msgfParTda-method}}
}

