% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{chargeRange}
\alias{chargeRange}
\alias{chargeRange<-}
\alias{chargeRange,msgfPar-method}
\alias{chargeRange<-,msgfPar,numeric-method}
\alias{chargeRange<-,msgfPar,msgfParChargeRange-method}
\title{Get and set the charge range in msgfPar objects}
\usage{
chargeRange(object)

chargeRange(object) <- value

\S4method{chargeRange}{msgfPar}(object)

\S4method{chargeRange}{msgfPar,numeric}(object) <- value

\S4method{chargeRange}{msgfPar,msgfParChargeRange}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either a numeric vector of length 2 or an msgfParChargeRange
object}
}
\value{
In case of the getter a numeric vector with the named elements 'min'
and 'max'
}
\description{
These functions allow you to retrieve and set the charge range in the msgfPar 
object of interest
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the charge range

\item \code{object = msgfPar,value = numeric}: Set the charge range using lower and upper bounds

\item \code{object = msgfPar,value = msgfParChargeRange}: Set the charge range using a dedicated
msgfParChargeRange object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
chargeRange(parameters) <- c(2, 4)
chargeRange(parameters)

}
\seealso{
Other msgfPar-getter_setter: \code{\link{db}},
  \code{\link{enzyme}}, \code{\link{fragmentation}},
  \code{\link{instrument}}, \code{\link{isotopeError}},
  \code{\link{lengthRange}}, \code{\link{matches}},
  \code{\link{mods}}, \code{\link{ntt}},
  \code{\link{protocol}}, \code{\link{tda}},
  \code{\link{tolerance}}
}
