% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{enzyme}
\alias{enzyme}
\alias{enzyme<-}
\alias{enzyme,msgfPar-method}
\alias{enzyme<-,msgfPar,numeric-method}
\alias{enzyme<-,msgfPar,character-method}
\alias{enzyme<-,msgfPar,msgfParEnzyme-method}
\title{Get and set enzyme in msgfPar objects}
\usage{
enzyme(object)

enzyme(object) <- value

\S4method{enzyme}{msgfPar}(object)

\S4method{enzyme}{msgfPar,numeric}(object) <- value

\S4method{enzyme}{msgfPar,character}(object) <- value

\S4method{enzyme}{msgfPar,msgfParEnzyme}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either an integer, string or msgfParEnzyme object}
}
\value{
In case of the getter a named integer
}
\description{
These functions allow you to retrieve and set the enzyme used for digestion
during sample treatment.
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the enzyme currently used

\item \code{object = msgfPar,value = numeric}: Set the enzyme to use using the key for the enzyme

\item \code{object = msgfPar,value = character}: Set the enzyme to use using the name of the enzyme

\item \code{object = msgfPar,value = msgfParEnzyme}: Set the enzyme to use using an msgfParEnzyme object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
enzyme(parameters) <- 'Trypsin'
enzyme(parameters) <- 3
enzyme(parameters)

}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{fragmentation}},
  \code{\link{instrument}}, \code{\link{isotopeError}},
  \code{\link{lengthRange}}, \code{\link{matches}},
  \code{\link{mods}}, \code{\link{ntt}},
  \code{\link{protocol}}, \code{\link{tda}},
  \code{\link{tolerance}}
}
