% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{fragmentation}
\alias{fragmentation}
\alias{fragmentation<-}
\alias{fragmentation,msgfPar-method}
\alias{fragmentation<-,msgfPar,numeric-method}
\alias{fragmentation<-,msgfPar,character-method}
\alias{fragmentation<-,msgfPar,msgfParFragmentation-method}
\title{Get and set fragmentation in msgfPar objects}
\usage{
fragmentation(object)

fragmentation(object) <- value

\S4method{fragmentation}{msgfPar}(object)

\S4method{fragmentation}{msgfPar,numeric}(object) <- value

\S4method{fragmentation}{msgfPar,character}(object) <- value

\S4method{fragmentation}{msgfPar,msgfParFragmentation}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either an integer, string or msgfParFragmentation object}
}
\value{
In case of the getter a named integer
}
\description{
These functions allow you to retrieve and set the fragmentation method used
during acquisition
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the fragmentation method currently used

\item \code{object = msgfPar,value = numeric}: Set the fragmentation method using the key for the 
method

\item \code{object = msgfPar,value = character}: Set the fragmentation method using the name of the 
method

\item \code{object = msgfPar,value = msgfParFragmentation}: Set the fragmentation method using an 
msgfParFragmentation object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
fragmentation(parameters) <- 'CID'
fragmentation(parameters) <- 3
fragmentation(parameters)

}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{enzyme}},
  \code{\link{instrument}}, \code{\link{isotopeError}},
  \code{\link{lengthRange}}, \code{\link{matches}},
  \code{\link{mods}}, \code{\link{ntt}},
  \code{\link{protocol}}, \code{\link{tda}},
  \code{\link{tolerance}}
}
