% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{isotopeError}
\alias{isotopeError}
\alias{isotopeError,msgfPar-method}
\alias{isotopeError<-}
\alias{isotopeError<-,msgfPar,msgfParIsotopeError-method}
\alias{isotopeError<-,msgfPar,numeric-method}
\title{Get and set isotope error in msgfPar objects}
\usage{
isotopeError(object)

isotopeError(object) <- value

\S4method{isotopeError}{msgfPar}(object)

\S4method{isotopeError}{msgfPar,numeric}(object) <- value

\S4method{isotopeError}{msgfPar,msgfParIsotopeError}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either an integer vector or an msgfParIsotopeError object}
}
\value{
In case of the getter an integer vector
}
\description{
These functions allow you to retrieve and set the isotope error used when
calculating the parent ion error range
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the isotope error currently used

\item \code{object = msgfPar,value = numeric}: Set the isotope error with an integer vector

\item \code{object = msgfPar,value = msgfParIsotopeError}: Set the isotope error with an msgfParIsotopeError
object
}}
\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
isotopeError(parameters) <- c(0, 3)
isotopeError(parameters)
}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{chargeRange,msgfPar-method}},
  \code{\link{chargeRange<-}},
  \code{\link{chargeRange<-,msgfPar,msgfParChargeRange-method}},
  \code{\link{chargeRange<-,msgfPar,numeric-method}};
  \code{\link{db}}, \code{\link{db,msgfPar-method}},
  \code{\link{db<-}},
  \code{\link{db<-,msgfPar,character-method}};
  \code{\link{enzyme}},
  \code{\link{enzyme,msgfPar-method}},
  \code{\link{enzyme<-}},
  \code{\link{enzyme<-,msgfPar,character-method}},
  \code{\link{enzyme<-,msgfPar,msgfParEnzyme-method}},
  \code{\link{enzyme<-,msgfPar,numeric-method}};
  \code{\link{fragmentation}},
  \code{\link{fragmentation,msgfPar-method}},
  \code{\link{fragmentation<-}},
  \code{\link{fragmentation<-,msgfPar,character-method}},
  \code{\link{fragmentation<-,msgfPar,msgfParFragmentation-method}},
  \code{\link{fragmentation<-,msgfPar,numeric-method}};
  \code{\link{instrument}},
  \code{\link{instrument,msgfPar-method}},
  \code{\link{instrument<-}},
  \code{\link{instrument<-,msgfPar,character-method}},
  \code{\link{instrument<-,msgfPar,msgfParInstrument-method}},
  \code{\link{instrument<-,msgfPar,numeric-method}};
  \code{\link{lengthRange}},
  \code{\link{lengthRange,msgfPar-method}},
  \code{\link{lengthRange<-}},
  \code{\link{lengthRange<-,msgfPar,msgfParLengthRange-method}},
  \code{\link{lengthRange<-,msgfPar,numeric-method}};
  \code{\link{matches}},
  \code{\link{matches,msgfPar-method}},
  \code{\link{matches<-}},
  \code{\link{matches<-,msgfPar,msgfParMatches-method}},
  \code{\link{matches<-,msgfPar,numeric-method}};
  \code{\link{mods}}, \code{\link{mods,msgfPar-method}},
  \code{\link{mods<-}},
  \code{\link{mods<-,msgfPar,msgfParModificationList-method}},
  \code{\link{nMod}}, \code{\link{nMod,msgfPar-method}},
  \code{\link{nMod<-}},
  \code{\link{nMod<-,msgfPar,numeric-method}};
  \code{\link{ntt}}, \code{\link{ntt,msgfPar-method}},
  \code{\link{ntt<-}},
  \code{\link{ntt<-,msgfPar,msgfParNtt-method}},
  \code{\link{ntt<-,msgfPar,numeric-method}};
  \code{\link{protocol}},
  \code{\link{protocol,msgfPar-method}},
  \code{\link{protocol<-}},
  \code{\link{protocol<-,msgfPar,character-method}},
  \code{\link{protocol<-,msgfPar,msgfParProtocol-method}},
  \code{\link{protocol<-,msgfPar,numeric-method}};
  \code{\link{tda}}, \code{\link{tda,msgfPar-method}},
  \code{\link{tda<-}},
  \code{\link{tda<-,msgfPar,logical-method}},
  \code{\link{tda<-,msgfPar,msgfParTda-method}};
  \code{\link{tolerance}},
  \code{\link{tolerance,msgfPar-method}},
  \code{\link{tolerance<-}},
  \code{\link{tolerance<-,msgfPar,character-method}},
  \code{\link{tolerance<-,msgfPar,msgfParTolerance-method}},
  \code{\link{toleranceRange}},
  \code{\link{toleranceRange,msgfPar-method}},
  \code{\link{toleranceRange<-}},
  \code{\link{toleranceRange<-,msgfPar,numeric-method}},
  \code{\link{toleranceUnit}},
  \code{\link{toleranceUnit,msgfPar-method}},
  \code{\link{toleranceUnit<-}},
  \code{\link{toleranceUnit<-,msgfPar,character-method}}
}

