% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{matches}
\alias{matches}
\alias{matches<-}
\alias{matches,msgfPar-method}
\alias{matches<-,msgfPar,numeric-method}
\alias{matches<-,msgfPar,msgfParMatches-method}
\title{Get and set the number of matches in msgfPar objects}
\usage{
matches(object)

matches(object) <- value

\S4method{matches}{msgfPar}(object)

\S4method{matches}{msgfPar,numeric}(object) <- value

\S4method{matches}{msgfPar,msgfParMatches}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either an integer or an msgfParMatches object}
}
\value{
In case of the getter an integer
}
\description{
These functions allow you to retrieve and set the number of matches per 
spectrum returned by MS-GF+
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the number of matches reported per spectrum

\item \code{object = msgfPar,value = numeric}: Set the number of matches reported per spectrum using an
integer

\item \code{object = msgfPar,value = msgfParMatches}: Set the number of matches reported per spectrum using an
msgfParMatches object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
matches(parameters) <- 5
matches(parameters)
}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{enzyme}},
  \code{\link{fragmentation}}, \code{\link{instrument}},
  \code{\link{isotopeError}}, \code{\link{lengthRange}},
  \code{\link{mods}}, \code{\link{ntt}},
  \code{\link{protocol}}, \code{\link{tda}},
  \code{\link{tolerance}}
}
