% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{msgfParChargeRange-class}
\alias{msgfParChargeRange}
\alias{msgfParChargeRange-class}
\title{A class handling charge ranges}
\usage{
msgfParChargeRange(value)
}
\arguments{
\item{value}{A numeric vector of length 2. The first element must be smaller than the last}
}
\value{
An msgfParChargeRange object
}
\description{
This class defines a charge range and provides methods to get correct system
call parameters.
}
\section{Slots}{

\describe{
 \item{\code{value}:}{A numeric vector of length 2 describing the upper and lower bounds of the charge range}
}
}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}};
  \code{\link{msgfParEnzyme}}; \code{\link{msgfPar}}
}

