% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParChargeRange.R
\docType{class}
\name{msgfParChargeRange-class}
\alias{msgfParChargeRange-class}
\alias{show,msgfParChargeRange-method}
\alias{length,msgfParChargeRange-method}
\alias{getMSGFpar,msgfParChargeRange-method}
\alias{msgfParChargeRange}
\title{A class handling charge ranges}
\usage{
\S4method{show}{msgfParChargeRange}(object)

\S4method{length}{msgfParChargeRange}(x)

\S4method{getMSGFpar}{msgfParChargeRange}(object)

msgfParChargeRange(value)
}
\arguments{
\item{object}{An msgfParChargeRange object}

\item{x}{An msgfParChargeRange object}

\item{value}{A numeric vector of length 2. The first element must be smaller 
than the last}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParChargeRange() An msgfParChargeRange object.
}
\description{
This class defines a charge range and provides methods to get correct system
call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParChargeRange object

\item \code{length}: Report the length of an msgfParChargeRange 
object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant 
function call
}}

\section{Slots}{

\describe{
\item{\code{value}}{A numeric vector of length 2 describing the upper and lower 
bounds of the charge range}
}}

\examples{
charges <- msgfParChargeRange(c(2, 5))

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
