% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParEnzyme.R
\docType{class}
\name{msgfParEnzyme-class}
\alias{msgfParEnzyme-class}
\alias{show,msgfParEnzyme-method}
\alias{length,msgfParEnzyme-method}
\alias{getMSGFpar,msgfParEnzyme-method}
\alias{msgfParEnzyme}
\title{A class handling enzyme selection}
\usage{
\S4method{show}{msgfParEnzyme}(object)

\S4method{length}{msgfParEnzyme}(x)

\S4method{getMSGFpar}{msgfParEnzyme}(object)

msgfParEnzyme(enzyme)
}
\arguments{
\item{object}{An msgfParEnzyme object}

\item{x}{An msgfParEnzyme object}

\item{enzyme}{Either an integer or a string}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParEnzyme() An msgfParEnzyme object.
}
\description{
This class defines a digestion enzyme selection and provides methods to get 
correct system call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParEnzyme object

\item \code{length}: Report the length of an msgfParEnzyme object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant function 
call
}}

\section{Slots}{

\describe{
\item{\code{enzyme}}{An integer specifiying the selection of enzyme. See the detail 
section of \code{\link{msgfPar}}}
}}

\examples{
enzyme <- msgfParEnzyme(1)
enzyme <- msgfParEnzyme('Trypsin')

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
