% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParInstrument.R
\docType{class}
\name{msgfParInstrument-class}
\alias{msgfParInstrument-class}
\alias{show,msgfParInstrument-method}
\alias{length,msgfParInstrument-method}
\alias{getMSGFpar,msgfParInstrument-method}
\alias{msgfParInstrument}
\title{A class handling instrument types}
\usage{
\S4method{show}{msgfParInstrument}(object)

\S4method{length}{msgfParInstrument}(x)

\S4method{getMSGFpar}{msgfParInstrument}(object)

msgfParInstrument(instrument)
}
\arguments{
\item{object}{An msgfParInstrument object}

\item{x}{An msgfParInstrument object}

\item{instrument}{An integer specifying the instrument type}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParInstrument() An msgfParInstrument object.
}
\description{
This class defines an instrument type and provides methods to get correct 
system
call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParInstrument object

\item \code{length}: Report the length of an msgfParInstrument 
object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant 
function call
}}

\section{Slots}{

\describe{
\item{\code{instrument}}{An integer specifying the instrument type}
}}

\examples{
instrument <- msgfParInstrument(1)
instrument <- msgfParInstrument('HighRes')

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
