% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParLengthRange.R
\docType{class}
\name{msgfParLengthRange-class}
\alias{msgfParLengthRange-class}
\alias{show,msgfParLengthRange-method}
\alias{length,msgfParLengthRange-method}
\alias{getMSGFpar,msgfParLengthRange-method}
\alias{msgfParLengthRange}
\title{A class handling length ranges}
\usage{
\S4method{show}{msgfParLengthRange}(object)

\S4method{length}{msgfParLengthRange}(x)

\S4method{getMSGFpar}{msgfParLengthRange}(object)

msgfParLengthRange(value)
}
\arguments{
\item{object}{An msgfParLengthRange object}

\item{x}{An msgfParLengthRange object}

\item{value}{A numeric vector of length 2. The first element must be smaller 
than the last}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParLengthRange() An msgfParLengthRange object.
}
\description{
This class defines a length range and provides methods to get correct system
call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParLengthRange object

\item \code{length}: Report the length of an msgfParLengthRange 
object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant 
function call
}}

\section{Slots}{

\describe{
\item{\code{value}}{A numeric vector of length 2 describing the upper and lower 
bounds of the length range}
}}

\examples{
lengths <- msgfParLengthRange(c(6, 40))

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
