% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParMatches.R
\docType{class}
\name{msgfParMatches-class}
\alias{msgfParMatches-class}
\alias{show,msgfParMatches-method}
\alias{length,msgfParMatches-method}
\alias{getMSGFpar,msgfParMatches-method}
\alias{msgfParMatches}
\title{A class handling number of matches}
\usage{
\S4method{show}{msgfParMatches}(object)

\S4method{length}{msgfParMatches}(x)

\S4method{getMSGFpar}{msgfParMatches}(object)

msgfParMatches(value)
}
\arguments{
\item{object}{An msgfParMatches object}

\item{x}{An msgfParMatches object}

\item{value}{An integer giving the number of matches that should be returned 
per spectrum}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParMatches() An msgfParMatches object.
}
\description{
This class defines a number of matches and provides methods to get correct 
system
call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParMatches object

\item \code{length}: Report the length of an msgfParMatches object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant function 
call
}}

\section{Slots}{

\describe{
\item{\code{value}}{An integer giving the number of matches per spectrum reported by 
MS-GF+}
}}

\examples{
matches <- msgfParMatches(5)

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
