% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParModification.R
\docType{class}
\name{msgfParModification-class}
\alias{msgfParModification-class}
\alias{getMSGFpar,msgfParModification-method}
\alias{show,msgfParModification-method}
\alias{msgfParModification}
\title{A class handling modification characteristics}
\usage{
\S4method{getMSGFpar}{msgfParModification}(object)

\S4method{show}{msgfParModification}(object)

msgfParModification(name, composition = "", mass = as.numeric(NA), residues,
  type, position)
}
\arguments{
\item{object}{An msgfParModification object}

\item{name}{The name of the modification}

\item{composition}{The molecular formular as a string for the modification. 
Loss of atoms are denoted with negative integers (e.g. O-1 for loss of 
oxygen)}

\item{mass}{The monoisotopic mass change between a peptide without and with
the given modification. Either composition or mass must be defined.}

\item{residues}{The amino acids that the modification applies to. Given as 
their one-letter code in upper-case without any separation. Use '*' for all
residues}

\item{type}{Either 'fix' or 'opt' for fixed or optional}

\item{position}{Where the modification can be. Either 'any', 'c-term',
'n-term', 'prot-c-term' or 'prot-n-term.}
}
\value{
For getMSGFpar() A string.

For msgfParModification() An msgfParModification object.
}
\description{
This class defines a single modification type that can be searched for by 
MS-GF+. Modifications are collected in a 
\code{\link{msgfParModificationList}} before adding them to 
\code{\linkS4class{msgfPar}} objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant 
function call

\item \code{show}: Short summary of msgfParModification object
}}

\section{Slots}{

\describe{
\item{\code{composition}}{The molecular formula for the modification.}

\item{\code{mass}}{The monoisotopic mass of the modification}

\item{\code{residues}}{The amino acids the modification applies to}

\item{\code{type}}{Whether the modification is optional or always present}

\item{\code{position}}{The possibel position of the modification}

\item{\code{name}}{The name of the modification}
}}

\examples{
# Using composition
modification1 <- msgfParModification(
                                     name='Carbamidomethyl',
                                     composition='C2H3N1O1',
                                     residues='C',
                                     type='fix',
                                     position='any'
                                    )
# Using exact mass
modification2 <- msgfParModification(
                                     name='Oxidation',
                                     mass=15.994915,
                                     residues='M',
                                     type='opt',
                                     position='any'
                                    )

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
