% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParModificationList.R
\docType{class}
\name{msgfParModificationList-class}
\alias{msgfParModificationList-class}
\alias{show,msgfParModificationList-method}
\alias{length,msgfParModificationList-method}
\alias{getMSGFpar,msgfParModificationList-method}
\alias{[[,msgfParModificationList,numeric,missing-method}
\alias{[[<-,msgfParModificationList,numeric,missing,msgfParModification-method}
\alias{msgfParModificationList}
\title{A class handling a list of modifications}
\usage{
\S4method{show}{msgfParModificationList}(object)

\S4method{length}{msgfParModificationList}(x)

\S4method{getMSGFpar}{msgfParModificationList}(object)

\S4method{[[}{msgfParModificationList,numeric,missing}(x, i, j, ...)

\S4method{[[}{msgfParModificationList,numeric,missing,msgfParModification}(x, i,
  j, ...) <- value

msgfParModificationList(nMod, modifications = list())
}
\arguments{
\item{object}{An msgfParModificationList object}

\item{x}{An msgfParModificationList object}

\item{i}{The index of the modification}

\item{j}{Ignored}

\item{...}{Ignored}

\item{value}{An msgfParModification object}

\item{nMod}{The maximum allowed number of modifications to expect on any 
peptide}

\item{modifications}{A list of \code{\linkS4class{msgfParModification}} 
objects}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For '[[' A msgfParModification object

For msgfParModificationList() An msgfParModificationList object.
}
\description{
This class defines a set of modifications and a maximum number of 
modifications allowed per peptide.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParModificationList 
object

\item \code{length}: Report the length of an 
msgfParModificationList object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant 
function call

\item \code{[[}: Get the i'th modification

\item \code{[[<-}: Set or change the i'th modification
}}

\section{Slots}{

\describe{
\item{\code{nMod}}{The maximum allowed number of modifications to expect on any 
peptide}

\item{\code{modifications}}{A list of \code{\linkS4class{msgfParModification}} 
objects}
}}

\examples{
modification1 <- msgfParModification(
                                     name='Carbamidomethyl',
                                     composition='C2H3N1O1',
                                     residues='C',
                                     type='fix',
                                     position='any'
                                    )
modification2 <- msgfParModification(
                                     name='Oxidation',
                                     mass=15.994915,
                                     residues='M',
                                     type='opt',
                                     position='any'
                                    )
modificationlist <- msgfParModificationList(
                                            nMod=2,
                                            modifications=list(
                                                modification1,
                                                modification2
                                            )
                                           )
modificationlist[[3]] <- msgfParModification(
                                             name='Gln->pyro-Glu',
                                             composition='H-3N-1',
                                             residues='Q',
                                             type='opt',
                                             position='N-term'
                                            )
}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
