% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParNtt.R
\docType{class}
\name{msgfParNtt-class}
\alias{msgfParNtt-class}
\alias{show,msgfParNtt-method}
\alias{length,msgfParNtt-method}
\alias{getMSGFpar,msgfParNtt-method}
\alias{msgfParNtt}
\title{A class handling cleavage specificity}
\usage{
\S4method{show}{msgfParNtt}(object)

\S4method{length}{msgfParNtt}(x)

\S4method{getMSGFpar}{msgfParNtt}(object)

msgfParNtt(value)
}
\arguments{
\item{object}{An msgfParNtt object}

\item{x}{An msgfParNtt object}

\item{value}{An integer between 0 and 2 that specifies the specificity (2: 
full cleavage, 1: semi specific cleavage, 0: no specificity)}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParNtt() An msgfParNtt object.
}
\description{
This class defines cleavage specificity and provides methods to get correct 
system call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParNtt object

\item \code{length}: Report the length of an msgfParNtt object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant function call
}}

\section{Slots}{

\describe{
\item{\code{value}}{An integer between 0 and 2 that specifies the specificity (2: 
full cleavage, 1: semi specific cleavage, 0: no specificity)}
}}

\examples{
ntt <- msgfParNtt(2)

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
