% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParProtocol.R
\docType{class}
\name{msgfParProtocol-class}
\alias{msgfParProtocol-class}
\alias{show,msgfParProtocol-method}
\alias{length,msgfParProtocol-method}
\alias{getMSGFpar,msgfParProtocol-method}
\alias{msgfParProtocol}
\title{A class handling protocol choice}
\usage{
\S4method{show}{msgfParProtocol}(object)

\S4method{length}{msgfParProtocol}(x)

\S4method{getMSGFpar}{msgfParProtocol}(object)

msgfParProtocol(protocol)
}
\arguments{
\item{object}{An msgfParProtocol object}

\item{x}{An msgfParProtocol object}

\item{protocol}{An integer or string specifying the protocol to use.}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParProtocol() An msgfParProtocol object.
}
\description{
This class defines a protocol and provides methods to get correct system
call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParProtocol object

\item \code{length}: Report the length of an msgfParProtocol object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant 
function call
}}

\section{Slots}{

\describe{
\item{\code{protocol}}{An integer specifying a specific protocol type}
}}

\examples{
protocol <- msgfParProtocol(0)
protocol <- msgfParProtocol('No protocol')

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParTda-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
