% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParTda.R
\docType{class}
\name{msgfParTda-class}
\alias{msgfParTda-class}
\alias{show,msgfParTda-method}
\alias{length,msgfParTda-method}
\alias{getMSGFpar,msgfParTda-method}
\alias{msgfParTda}
\title{A class handling use of target-decoy approach for FDR estimation}
\usage{
\S4method{show}{msgfParTda}(object)

\S4method{length}{msgfParTda}(x)

\S4method{getMSGFpar}{msgfParTda}(object)

msgfParTda(value)
}
\arguments{
\item{object}{An msgfParTda object}

\item{x}{An msgfParTda object}

\item{value}{A boolean defining whether to use tda or not}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParTda() An msgfParTda object.
}
\description{
This class defines whether to use target-decoy approach and provides methods 
to get correct system call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParTda object

\item \code{length}: Report the length of an msgfParTda object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant function call
}}

\section{Slots}{

\describe{
\item{\code{tda}}{A boolean defining whether to use tda or not}
}}

\examples{
tda <- msgfParTda(TRUE)

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTolerance-class}},
  \code{\link{msgfPar}}
}
