% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgfParTolerance.R
\docType{class}
\name{msgfParTolerance-class}
\alias{msgfParTolerance-class}
\alias{show,msgfParTolerance-method}
\alias{length,msgfParTolerance-method}
\alias{getMSGFpar,msgfParTolerance-method}
\alias{msgfParTolerance}
\title{A class handling parent ion tolerance}
\usage{
\S4method{show}{msgfParTolerance}(object)

\S4method{length}{msgfParTolerance}(x)

\S4method{getMSGFpar}{msgfParTolerance}(object)

msgfParTolerance(value, low, high, unit)
}
\arguments{
\item{object}{An msgfParTolerance object}

\item{x}{An msgfParTolerance object}

\item{value}{A numeric giving the upper and lower bounds of the tolerance}

\item{low}{A numeric giving the lower bounds of the tolerance}

\item{high}{A numeric giving the higher bounds of the tolerance}

\item{unit}{The unit used. Either 'ppm' or 'Da'}
}
\value{
For length() An integer.

For getMSGFpar() A string.

For msgfParTolerance() An msgfParTolerance object.
}
\description{
This class defines a parent ion tolerance and provides methods to get correct 
system call parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of msgfParTolerance object

\item \code{length}: Report the length of an msgfParTolerance object

\item \code{getMSGFpar}: Get \code{\link[base]{system}} compliant 
function call
}}

\section{Slots}{

\describe{
\item{\code{unit}}{The unit used to define the tolerance}

\item{\code{low}}{The lower bound of the tolerance}

\item{\code{high}}{The higher bound of the tolerance}
}}

\examples{
# Symmetric
tolerance <- msgfParTolerance(20, unit='ppm')

# Asymmetric
tolerance <- msgfParTolerance(low=0.5, high=1.5, unit='Da')

}
\seealso{
Other msgfParClasses: \code{\link{msgfPar-class}},
  \code{\link{msgfParChargeRange-class}},
  \code{\link{msgfParEnzyme-class}},
  \code{\link{msgfParFragmentation-class}},
  \code{\link{msgfParInstrument-class}},
  \code{\link{msgfParIsotopeError-class}},
  \code{\link{msgfParLengthRange-class}},
  \code{\link{msgfParMatches-class}},
  \code{\link{msgfParModification-class}},
  \code{\link{msgfParModificationList-class}},
  \code{\link{msgfParNtt-class}},
  \code{\link{msgfParProtocol-class}},
  \code{\link{msgfParTda-class}}, \code{\link{msgfPar}}
}
