% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGFplus-package.R
\docType{package}
\name{MSGFplus-package}
\alias{MSGFplus-package}
\title{An interface between R and MS-GF+}
\description{
This package comes bundled with the peptide database search tool MS-GF+ and
provides means of running peptide searches from within R with automatic
result import and easy handling of parameters.
}
\details{
The version bundled with this package is: MS-GF+ Beta (v10072) (11/11/2013)
}
\author{
Thomas Lin Pedersen with contributions from Laurent Gatto
}
