% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{ntt}
\alias{ntt}
\alias{ntt<-}
\alias{ntt,msgfPar-method}
\alias{ntt<-,msgfPar,numeric-method}
\alias{ntt<-,msgfPar,msgfParNtt-method}
\title{Get and set cleavage specificity in msgfPar objects}
\usage{
ntt(object)

ntt(object) <- value

\S4method{ntt}{msgfPar}(object)

\S4method{ntt}{msgfPar,numeric}(object) <- value

\S4method{ntt}{msgfPar,msgfParNtt}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{An integer or an msgfParNtt object}
}
\value{
In case of the getter an integer between 0 and 2
}
\description{
These functions allow you to retrieve and set the quality of cleavage allowed
during search in MS-GF+ (number of tolerable termini - ntt)
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the number of tolerable termini

\item \code{object = msgfPar,value = numeric}: Set the ntt using an integer

\item \code{object = msgfPar,value = msgfParNtt}: Set the ntt using an msgfParNtt object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
ntt(parameters) <- 2
ntt(parameters)

}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{enzyme}},
  \code{\link{fragmentation}}, \code{\link{instrument}},
  \code{\link{isotopeError}}, \code{\link{lengthRange}},
  \code{\link{matches}}, \code{\link{mods}},
  \code{\link{protocol}}, \code{\link{tda}},
  \code{\link{tolerance}}
}
