% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{protocol}
\alias{protocol}
\alias{protocol<-}
\alias{protocol,msgfPar-method}
\alias{protocol<-,msgfPar,numeric-method}
\alias{protocol<-,msgfPar,character-method}
\alias{protocol<-,msgfPar,msgfParProtocol-method}
\title{Get and set protocol in msgfPar objects}
\usage{
protocol(object)

protocol(object) <- value

\S4method{protocol}{msgfPar}(object)

\S4method{protocol}{msgfPar,numeric}(object) <- value

\S4method{protocol}{msgfPar,character}(object) <- value

\S4method{protocol}{msgfPar,msgfParProtocol}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either an integer, a string or an msgfParProtocol object}
}
\value{
In case of the getter a named integer
}
\description{
These functions allow you to retrieve and set the protocol used during MS-GF+
analysis. This allows you to fine tune the analysis in case of labelled or
phosphoproteomic analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the protocol currently used

\item \code{object = msgfPar,value = numeric}: Set the protocol using the key for the protocol

\item \code{object = msgfPar,value = character}: Set the protocol using the name of the protocol

\item \code{object = msgfPar,value = msgfParProtocol}: Set the protocol using an msgfParProtocol object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
protocol(parameters) <- 'Phosphorylation'
protocol(parameters) <- 0
protocol(parameters)

}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{enzyme}},
  \code{\link{fragmentation}}, \code{\link{instrument}},
  \code{\link{isotopeError}}, \code{\link{lengthRange}},
  \code{\link{matches}}, \code{\link{mods}},
  \code{\link{ntt}}, \code{\link{tda}},
  \code{\link{tolerance}}
}
