% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{tda}
\alias{tda}
\alias{tda<-}
\alias{tda,msgfPar-method}
\alias{tda<-,msgfPar,logical-method}
\alias{tda<-,msgfPar,msgfParTda-method}
\title{Get and set use of target-decoy approach in msgfPar objects}
\usage{
tda(object)

tda(object) <- value

\S4method{tda}{msgfPar}(object)

\S4method{tda}{msgfPar,logical}(object) <- value

\S4method{tda}{msgfPar,msgfParTda}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{Either a boolean or msgfParTda object}
}
\value{
In case of the getter a boolean indicating whether tda is used or not
}
\description{
These functions allow you to retrieve and set whether the target-decoy 
approach should be used to estimate q-values.
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get whether tda is currently used for FDR estimation

\item \code{object = msgfPar,value = logical}: Set the use of tda using a boolean (TRUE/FALSE)

\item \code{object = msgfPar,value = msgfParTda}: Set the use of tda using an msgfParTda object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
tda(parameters) <- TRUE
tda(parameters)

}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{enzyme}},
  \code{\link{fragmentation}}, \code{\link{instrument}},
  \code{\link{isotopeError}}, \code{\link{lengthRange}},
  \code{\link{matches}}, \code{\link{mods}},
  \code{\link{ntt}}, \code{\link{protocol}},
  \code{\link{tolerance}}
}
