% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/msgfPar-getters.R
\docType{methods}
\name{tolerance}
\alias{tolerance}
\alias{tolerance<-}
\alias{toleranceRange}
\alias{toleranceRange<-}
\alias{toleranceUnit}
\alias{toleranceUnit<-}
\alias{tolerance,msgfPar-method}
\alias{toleranceRange,msgfPar-method}
\alias{toleranceUnit,msgfPar-method}
\alias{toleranceRange<-,msgfPar,numeric-method}
\alias{toleranceUnit<-,msgfPar,character-method}
\alias{tolerance<-,msgfPar,character-method}
\alias{tolerance<-,msgfPar,msgfParTolerance-method}
\title{Get and set the parent tolerance in msgfPar objects}
\usage{
tolerance(object)

tolerance(object) <- value

toleranceRange(object)

toleranceRange(object) <- value

toleranceUnit(object)

toleranceUnit(object) <- value

\S4method{tolerance}{msgfPar}(object)

\S4method{toleranceRange}{msgfPar}(object)

\S4method{toleranceUnit}{msgfPar}(object)

\S4method{toleranceRange}{msgfPar,numeric}(object) <- value

\S4method{toleranceUnit}{msgfPar,character}(object) <- value

\S4method{tolerance}{msgfPar,character}(object) <- value

\S4method{tolerance}{msgfPar,msgfParTolerance}(object) <- value
}
\arguments{
\item{object}{An msgfPar object}

\item{value}{For tolerance a character vector of length 2, each element of 
the form '<value> <unit>'. For toleranceUnit a string. For toleranceRange a 
numeric vector of length 2.}
}
\value{
For tolerance a character vector with the lower and upper tolerance 
limit with unit. For toleranceUnit a string with the unit used. For 
toleranceRange a numeric vector with lower and upper tolerance limit.
}
\description{
These functions allow you to retrieve and set the tolerance used for matching
parent ions to peptides in the database
}
\section{Methods (by class)}{
\itemize{
\item \code{msgfPar}: Get the lower and upper bounds of the tolerance

\item \code{msgfPar}: Get the lower and upper bounds of the tolerance

\item \code{msgfPar}: Get the unit the tolerance is measured in

\item \code{object = msgfPar,value = numeric}: Set the lower and upper bounds of the tolerance using a
numeric vector of length 2

\item \code{object = msgfPar,value = character}: Set the unit the tolerance is meassured in

\item \code{object = msgfPar,value = character}: Set the lower and upper bounds of the tolerance using a
character vector of length 2, where each element is of the form '<value> 
<unit>'

\item \code{object = msgfPar,value = msgfParTolerance}: Set the lower and upper bounds of the tolerance using 
an msgfParTolerance object
}}

\examples{
parameters <- msgfPar(system.file(package='MSGFplus', 'extdata', 'milk-proteins.fasta'))
tolerance(parameters) <- c('20 ppm', '20 ppm')
toleranceUnit(parameters) <- 'Da'
toleranceRange(parameters) <- c(1.5, 1.5)
tolerance(parameters)

}
\seealso{
Other msgfPar-getter_setter: \code{\link{chargeRange}},
  \code{\link{db}}, \code{\link{enzyme}},
  \code{\link{fragmentation}}, \code{\link{instrument}},
  \code{\link{isotopeError}}, \code{\link{lengthRange}},
  \code{\link{matches}}, \code{\link{mods}},
  \code{\link{ntt}}, \code{\link{protocol}},
  \code{\link{tda}}
}
